refactor: merge ftp-storage into ftp.service with PassThrough
- Move ftpStorage function to ftp.service.ts
- Use PassInstead of /tmp/ storage for hash calculation
- Remove ftp-storage.ts
- Update imports in common.controller.ts
